Extract available slot computation
